remove redundant AepInboxEventObserver implementation
