+++ code/messaging/src/main/java/com/adobe/marketing/mobile/messaging/InboxEventObserver.kt
@@ -1,5 +1,5 @@
 /*
-  Copyright 2025 Adobe. All rights reserved.
+  Copyright 2026 Adobe. All rights reserved.
   This file is licensed to you under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License. You may obtain a copy
   of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -13,21 +13,42 @@ package com.adobe.marketing.mobile.messaging
 
 import com.adobe.marketing.mobile.MessagingEdgeEventType
 import com.adobe.marketing.mobile.aepcomposeui.InboxEvent
+import com.adobe.marketing.mobile.aepcomposeui.UIEvent
+import com.adobe.marketing.mobile.aepcomposeui.observers.AepInboxEventObserver
+import com.adobe.marketing.mobile.aepcomposeui.observers.AepUIEventObserver
 
 /**
- * Handles tracking for inbox-level events. Used internally by [MessagingInboxEventObserver].
+ * Messaging implementation of [AepInboxEventObserver].
+ * Handles tracking for inbox, delegates state updates ([MessagingInboxProvider.onInboxEvent]), and
+ * forwards item-level events to the provided [AepUIEventObserver] or a default [ContentCardEventObserver] when null.
+ *
+ * @param provider The [MessagingInboxProvider] that owns inbox state; the observer calls
+ *   [MessagingInboxProvider.onInboxEvent] so the provider updates its state.
+ * @param itemEventObserver Optional [AepUIEventObserver] for item-level events (e.g. [ContentCardEventObserver]).
+ *   When null, a default [ContentCardEventObserver] with null callback is used.
  */
-internal class InboxEventObserver {
+class InboxEventObserver(
+    private val provider: MessagingInboxProvider,
+    private val itemEventObserver: AepUIEventObserver? = null
+) : AepInboxEventObserver {
 
-    /**
-     * Performs tracking for inbox events.
-     */
-    internal fun onInboxEvent(event: InboxEvent) {
+    private val observer: AepUIEventObserver by lazy {
+        itemEventObserver ?: ContentCardEventObserver(null)
+    }
+
+    override fun onInboxEvent(event: InboxEvent) {
         when (event) {
             is InboxEvent.Display -> {
-                ContentCardMapper.instance.getInboxPropositionItem(event.inboxUIState.template.id)
-                    ?.track(MessagingEdgeEventType.DISPLAY)
+                if (!event.inboxUIState.displayed) {
+                    ContentCardMapper.instance.getInboxPropositionItem(event.inboxUIState.template.id)
+                        ?.track(MessagingEdgeEventType.DISPLAY)
+                    provider.onInboxEvent(event)
+                }
             }
         }
+    }
+
+    override fun onEvent(event: UIEvent<*, *>) {
+        observer.onEvent(event)
     }
 }
